requirements.txt
arxiv
chainlit
python-dotenv
langchain-core
langchain-community
langchain
langchain_text_splitters
langchain-chroma
langchain-openai